Move focus model code to EnggUtils
